pyproject.toml
[tool.poetry]
name = "AGR-MVS"
version = "0.0.4"
description = "FRANCAPP Proyecto IV-22-23"
authors = ["Alvaro García Rodríguez <alvarogaro@correo.ugr.es>"]
license = "GPL-3.0-only"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pyhamcrest = "^2.0.4"
pytest = "^7.2.0"

# Automatización de Tareas

[tool.poe.tasks]
install = "poetry install"
check = "python -m compileall src"
test = "poetry run pytest -v tests/test-agr-mvs.py"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
